+++ src/sd_initialProducts.cu
@@ -33,7 +33,7 @@ string radiance_function(Products products, MTL mtl)
 
 string reflectance_function(Products products, MTL mtl)
 {
-    const float sin_sun = sin(mtl.sun_elevation * PI / 180);
+    const double sin_sun = sin(mtl.sun_elevation * PI / 180);
 
     int64_t initial_time, final_time;
     cudaEvent_t start, stop;
